script/retrival/dataset_retrieval/requirements.txt
torch>=1.9.0
transformers>=4.21.0
numpy>=1.21.0
tqdm>=4.62.0
faiss-cpu>=1.7.4
